+++ docs/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Full backup downloads now use the same compact hour/day usage rollups as Gist instead of serializing the entire raw scan cache.
+
 ## [1.0.5] - 2026-07-13
 
 ### Changed